Fix movorder  and new heuristic +70  bug fixes +400 elo
Bench 4866158

+++ src/Engine/Evaluation.java
@@ -66,7 +66,7 @@ public class Evaluation {
         }
 
         if (board.hasWinWithFurtherOffset(1, side)) {
-            eval += (short) (1000 - ply);
+            eval += 1000 - ply;
         }
         return eval;
     }
